renderdjangotest1-main/requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
dataclasses==0.6
Django==3.2.20
django-cors-headers==4.1.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
filelock==3.12.2
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
mpmath==1.3.0
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.8.0.74
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
protobuf==4.23.4
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pymongo==3.12.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.7.3
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
sqlparse==0.2.4
sympy==1.10.1
thop==0.1.1.post2209072238
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.143
uritemplate==4.1.1
urllib3==1.26.16
zipp==3.15.0
